Implemented workaround for SharpZipLib bug, this fixes #6.
Looks like the codepage detection fails on Windows 10 when the system and the input locale is not identical. This bug is reported in icsharpcode/SharpZipLib#195. The workaround is to set the codepage to 850 if the detection failed. That should work, because the file name in the ZIP is in ASCII-7 anyways.